requirements.txt
paho-mqtt==2.1.0
pyusb==1.2.1
